+++ localcues/gradient.h
@@ -1,10 +1,10 @@
 #ifndef GRADIENT_H
 #define GRADIENT_H
 
-int initializeGradients(uint width, uint height, uint border, uint maxbins, uint norients, uint nscale, uint textonChoice);
+int initializeGradients(unsigned long width, unsigned long height, unsigned long border, unsigned long maxbins, unsigned long norients, unsigned long nscale, unsigned long textonChoice);
 
-float* gradients(float* devImage, uint nbins, bool blur, float sigma, uint* radii, int textonChoice);
-float* gradients(int* devImage, uint nbins, bool blur, float sigma, uint* radii, int textonChoice);
+float* gradients(float* devImage, unsigned long nbins, bool blur, float sigma, unsigned long* radii, int textonChoice);
+float* gradients(int* devImage, unsigned long nbins, bool blur, float sigma, unsigned long* radii, int textonChoice);
 
 
 void finalizeGradients();
